Implement sort linking
This approach makes use of the subroutine jump command.

That is,

00 SYSCLIP
01 LCOORDS ----\
02 DRAWEND <-\ |
03 Polygon   | |
0i Polygon <-|-/
0n Polygon --/

After processing CMDT 01, a subroutine call is performed, jumping to the
last to draw polygon. From that polygon, the next polygon (via "jump
assign") is processed. All polygons are processed. The last polygon to
be processed calls to return from subroutine, bringing the command table
pointer back to CMDT 02, ending drawing.